requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-RESTful==0.3.10
marshmallow==3.20.1
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
flask-swagger-ui==4.11.1
apispec==6.3.0
apispec-webframeworks==0.5.2
pytest==7.4.0
python-dotenv==1.0.0
email-validator==2.0.0
